+++ src/acconeer/exptool/a121/algo/bilateration/__init__.py
@@ -1,4 +1,5 @@
 # Copyright (c) Acconeer AB, 2022
 # All rights reserved
 
+from ._configs import get_default_detector_config
 from ._processor import ProcessorConfig