+++ src/acconeer/exptool/a121/algo/distance/_detector_plugin.py
@@ -4,8 +4,9 @@
 from __future__ import annotations
 
 import logging
+from enum import Enum, auto
 from pathlib import Path
-from typing import Any, Callable, Optional
+from typing import Any, Callable, Mapping, Optional
 
 import attrs
 import numpy as np
@@ -33,6 +34,7 @@ from acconeer.exptool.app.new import (
     Message,
     PluginFamily,
     PluginGeneration,
+    PluginPresetBase,
     PluginSpecBase,
     PluginState,
     PluginStateMessage,
@@ -67,7 +69,16 @@ class SharedState:
     replaying: bool = attrs.field(default=False)
 
 
+class PluginPresetId(Enum):
+    DEFAULT = auto()
+
+
 class BackendPlugin(DetectorBackendPluginBase[SharedState]):
+
+    PLUGIN_PRESETS: Mapping[int, Callable[[], DetectorConfig]] = {
+        PluginPresetId.DEFAULT.value: lambda: DetectorConfig()
+    }
+
     def __init__(
         self, callback: Callable[[Message], None], generation: PluginGeneration, key: str
     ) -> None:
@@ -134,6 +145,12 @@ class BackendPlugin(DetectorBackendPluginBase[SharedState]):
     def update_sensor_ids(self, *, sensor_ids: list[int]) -> None:
         self.shared_state.sensor_ids = sensor_ids
         self.broadcast()
+
+    @is_task
+    def set_preset(self, preset_id: int) -> None:
+        preset_config = self.PLUGIN_PRESETS[preset_id]
+        self.shared_state.config = preset_config()
+        self.broadcast(sync=True)
 
     def teardown(self) -> None:
         try:
@@ -687,4 +704,8 @@ DISTANCE_DETECTOR_PLUGIN = PluginSpec(
     title="Distance detector",
     description="Easily measure distance to objects.",
     family=PluginFamily.DETECTOR,
+    presets=[
+        PluginPresetBase(name="Default", preset_id=PluginPresetId.DEFAULT),
+    ],
+    default_preset_id=PluginPresetId.DEFAULT,
 )