+++ src/acconeer/exptool/app/new/__init__.py
@@ -20,6 +20,6 @@ from .backend import (
     PluginStateMessage,
     is_task,
 )
-from .pluginbase import PlotPluginBase, PluginSpecBase, ViewPluginBase
+from .pluginbase import PlotPluginBase, PluginPresetBase, PluginSpecBase, ViewPluginBase
 from .storage import get_temp_dir, get_temp_h5_path
 from .ui import BUTTON_ICON_COLOR