+++ src/acconeer/exptool/app/new/app_model/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) Acconeer AB, 2022
 # All rights reserved
 
-from .app_model import AppModel, PluginSpec
+from .app_model import AppModel, PluginPresetSpec, PluginSpec
 from .app_model_listener import AppModelListener
 from .plugin_protocols import PlotPluginInterface, ViewPluginInterface