+++ src/acconeer/exptool/app/new/pluginbase/__init__.py
@@ -2,5 +2,6 @@
 # All rights reserved
 
 from .plot_plugin_base import PlotPluginBase
+from .plugin_preset_base import PluginPresetBase
 from .plugin_spec_base import PluginSpecBase
 from .view_plugin_base import ViewPluginBase